+++ docs/versionhistory.rst
@@ -3,7 +3,7 @@ Version history
 
 This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
-**UNRELEASED**
+**2.0.0rc2**
 
 - Changed the ``repr()`` of ``ExceptionGroup`` to match trio's ``MultiError``
 - Fixed ``TLSStream`` not relaying the extra attributes from its transport stream